fix: use booleans and not numbers
From @HinTak: This cannot be classified as a 'refactor" - it is a documented difference between sdl2 and sdl3 - the sdl3 versions returns booleans (False on success), the sdl2 returns zero on success. And in python, comparing True < 0 and False < 0 are both False. So your code is simply wrong in sdl3 as the bits after if will never been triggered, and your code would continue despite errors (if there is one) and bomb out later.